// src/containers/SignupForm/SignupForm.types.ts
/**
 * Defines the structure of signup form's data
 */
export interface SignupFormDataType {
    /**
     * Email field in signup form
     */
    email: string;

    /**
     * Name field in signup form
     */
    name: string;

    /**
     * Phone number field in signup form
     */
    phone: string;

    /**
     * Password field in signup form
     */
    password: string;

    /**
     * Confirm Password field in signup form
     */
    confirm_password: string;
}